refactor(user-table): spliting the large component to less dense files
